CB-23803 Delete monitor should be executed on Leader node only Currently, Quartz job is not running in cluster mode. With this, it's possible for multiple pods to execute the delete job concurrently. It will have no functional impact since cluster deletion will be successful only one one node and fail on others causing error logs to appear. We should enhance Periscope to execute delete job only on leader node since there is only one leader.

+++ autoscale/src/main/java/com/sequenceiq/periscope/monitor/DeleteMonitor.java
@@ -1,5 +1,6 @@
 package com.sequenceiq.periscope.monitor;
 
+import java.util.Optional;
 import java.util.concurrent.TimeUnit;
 
 import org.quartz.JobExecutionContext;
@@ -10,8 +11,11 @@ import org.springframework.stereotype.Component;
 
 import com.sequenceiq.periscope.domain.Cluster;
 import com.sequenceiq.periscope.domain.PeriscopeJob;
+import com.sequenceiq.periscope.domain.PeriscopeNode;
 import com.sequenceiq.periscope.monitor.handler.ClusterDeleteHandler;
 import com.sequenceiq.periscope.repository.PeriscopeJobRepository;
+import com.sequenceiq.periscope.repository.PeriscopeNodeRepository;
+import com.sequenceiq.periscope.service.ha.PeriscopeNodeConfig;
 
 @Component
 @ConditionalOnProperty(prefix = "periscope.enabledAutoscaleMonitors.delete-monitor", name = "enabled", havingValue = "true")
@@ -24,6 +28,10 @@ public class DeleteMonitor extends ClusterMonitor {
     private ClusterDeleteHandler clusterDeleteHandler;
 
     private PeriscopeJobRepository periscopeJobRepository;
+
+    private PeriscopeNodeRepository periscopeNodeRepository;
+
+    private PeriscopeNodeConfig periscopeNodeConfig;
 
     @Override
     public String getIdentifier() {
@@ -42,16 +50,31 @@ public class DeleteMonitor extends ClusterMonitor {
 
     @Override
     public void execute(JobExecutionContext context) {
-        LOGGER.info("Execution for {} started", getIdentifier());
-        long lastExecuted = System.currentTimeMillis();
         evalContext(context);
-        clusterDeleteHandler = getApplicationContext().getBean(ClusterDeleteHandler.class);
-        periscopeJobRepository = getApplicationContext().getBean(PeriscopeJobRepository.class);
-        PeriscopeJob periscopeJob = periscopeJobRepository.findById(getIdentifier()).orElse(new PeriscopeJob(getIdentifier(),
-                System.currentTimeMillis() - DEFAULT_DELETE_DURATION_MS));
-        clusterDeleteHandler.deleteClusters(periscopeJob.getLastExecuted());
-        periscopeJob.setLastExecuted(lastExecuted);
-        periscopeJobRepository.save(periscopeJob);
-        LOGGER.info("Execution for {} completed successfully with lastExecuted updated as {}", getIdentifier(), lastExecuted);
+        boolean leader = false;
+        periscopeNodeConfig = getApplicationContext().getBean(PeriscopeNodeConfig.class);
+        if (periscopeNodeConfig.getId() != null) {
+            periscopeNodeRepository = getApplicationContext().getBean(PeriscopeNodeRepository.class);
+            Optional<PeriscopeNode> periscopeNodeOptional = periscopeNodeRepository.findById(periscopeNodeConfig.getId());
+            if (periscopeNodeOptional.isPresent()) {
+                leader = periscopeNodeOptional.get().isLeader();
+            }
+            if (leader) {
+                LOGGER.info("Execution for {} started", getIdentifier());
+                long lastExecuted = System.currentTimeMillis();
+                clusterDeleteHandler = getApplicationContext().getBean(ClusterDeleteHandler.class);
+                periscopeJobRepository = getApplicationContext().getBean(PeriscopeJobRepository.class);
+                PeriscopeJob periscopeJob = periscopeJobRepository.findById(getIdentifier()).orElse(new PeriscopeJob(getIdentifier(),
+                        System.currentTimeMillis() - DEFAULT_DELETE_DURATION_MS));
+                clusterDeleteHandler.deleteClusters(periscopeJob.getLastExecuted());
+                periscopeJob.setLastExecuted(lastExecuted);
+                periscopeJobRepository.save(periscopeJob);
+                LOGGER.info("Execution for {} completed successfully with lastExecuted updated as {}", getIdentifier(), lastExecuted);
+            } else {
+                LOGGER.info("Node is not leader so execution of delete monitor will be skipped");
+            }
+        } else {
+            LOGGER.warn("Node Id is not specified so execution of delete monitor will be skipped");
+        }
     }
 }
